requirements.txt
numpy==2.0.1
pandas==2.2.2
scikit-learn==1.5.1
matplotlib==3.9.0
shap==0.45.1
typer==0.12.3
PyYAML==6.0.2
joblib==1.4.2

# dev tooling
pytest==8.2.0
pre-commit==3.8.0
black==24.8.0
ruff==0.6.4
